requirements.txt
streamlit
facebook-business
google-generativeai